LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/navigate-cross-document-double-expected.txt



FAIL event and promise ordering when navigate() is called to a cross-document destination, interrupting another navigate() to a cross-document destination assert_array_equals: expected property 2 to be "navigateerror" but got "navigate" (expected array ["navigate", "AbortSignal abort", "navigateerror", "navigate", "committed rejected 1", "finished rejected 1", "promise microtask"] got ["navigate", "AbortSignal abort", "navigate", "navigate", "AbortSignal abort", "navigate", "promise microtask"])
